+++ lib/src/database/accounting_database.dart
@@ -4,6 +4,7 @@
 import 'dart:io';
 import 'package:drift/drift.dart';
 import 'package:drift/native.dart';
+import 'package:flutter_accounting/flutter_accounting.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:path/path.dart' as p;